optimized authenticator entities
