+++ components/2FABundle/bundle/Entity/AuthenticatorInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * NovaeZ2FABundle.
+ *
+ * @package   NovaeZ2FABundle
+ *
+ * @author    Yassine HANINI
+ * @copyright 2021 AlmaviaCX
+ * @license   https://github.com/Novactive/NovaeZ2FA/blob/main/LICENSE
+ */
+
+declare(strict_types=1);
+
+namespace Novactive\Bundle\eZ2FABundle\Entity;
+
+interface AuthenticatorInterface
+{
+    public function setAuthenticatorSecret(?string $googleAuthenticatorSecret): void;
+}
