+++ components/2FABundle/bundle/Entity/UserGoogleAuthSecret.php
@@ -16,20 +16,16 @@ namespace Novactive\Bundle\eZ2FABundle\Entity;
 
 use eZ\Publish\API\Repository\Values\User\User as APIUser;
 use eZ\Publish\Core\MVC\Symfony\Security\User;
-use Scheb\TwoFactorBundle\Model\BackupCodeInterface;
 use Scheb\TwoFactorBundle\Model\Google\TwoFactorInterface;
 
-final class UserGoogleAuthSecret extends User implements TwoFactorInterface, BackupCodeInterface
+final class UserGoogleAuthSecret extends User implements TwoFactorInterface, BackupCodeInterface, AuthenticatorInterface
 {
+    use BackupCodeAware;
+
     /**
      * @var string|null
      */
     private $secret;
-
-    /**
-     * @var array
-     */
-    private $backupCodes = [];
 
     public function __construct(APIUser $user, array $roles = [], ?string $secret = null)
     {
@@ -56,29 +52,6 @@ final class UserGoogleAuthSecret extends User implements TwoFactorInterface, Bac
     public function setAuthenticatorSecret(?string $googleAuthenticatorSecret): void
     {
         $this->secret = $googleAuthenticatorSecret;
-    }
-
-    public function isBackupCode(string $code): bool
-    {
-        return in_array((int) $code, $this->backupCodes, true);
-    }
-
-    public function invalidateBackupCode(string $code): void
-    {
-        $key = array_search((int) $code, $this->backupCodes, true);
-        if (false !== $key) {
-            unset($this->backupCodes[$key]);
-        }
-    }
-
-    public function setBackupCodes(array $backupCodes): void
-    {
-        $this->backupCodes = $backupCodes;
-    }
-
-    public function getBackupCodes(): array
-    {
-        return $this->backupCodes;
     }
 
     public function __serialize(): array
